+++ lib/dato/fragments/meta_tags.rb
@@ -1,0 +1,5 @@
+Dato::Fragments::MetaTags = GQLi::DSL.fragment("metaTagsFragment", "Tag") {
+  attributes
+  content
+  tag
+}